+++ examples/index.js
@@ -8,6 +8,8 @@ const svgEl = document.getElementById('svg');
 const photosCreditEl = document.getElementById('photos-credit');
 const matricesCreditEl = document.getElementById('matrices-credit');
 const svgCreditEl = document.getElementById('svg-credit');
+const optionsEl = document.getElementById('options');
+const optionsTogglerEl = document.getElementById('options-toggler');
 const undoButton = document.getElementById('undo');
 
 let piling;
@@ -21,7 +23,7 @@ const undoHandler = () => {
   // Remove the current history
   history.pop();
   piling.importState(history[history.length - 1]);
-  if (history.length === 0) undoButton.style.display = 'none';
+  if (history.length === 0) undoButton.disabled = true;
 };
 
 undoButton.addEventListener('click', undoHandler);
@@ -35,7 +37,7 @@ const ignoredActions = new Set([
 const updateHandler = ({ action }) => {
   if (ignoredActions.has(action.type)) return;
 
-  undoButton.style.display = 'block';
+  undoButton.disabled = false;
 
   const state = piling.exportState();
   history.push(state);
@@ -55,7 +57,7 @@ const createPiles = async example => {
       svgCreditEl.style.display = 'none';
       photosEl.style.display = 'block';
       photosCreditEl.style.display = 'block';
-      undoButton.style.display = 'none';
+      undoButton.disabled = true;
       piling = await createPhotoPiles(photosEl);
       history = [];
       piling.subscribe('update', updateHandler);
@@ -69,7 +71,7 @@ const createPiles = async example => {
       svgCreditEl.style.display = 'none';
       matricesEl.style.display = 'block';
       matricesCreditEl.style.display = 'block';
-      undoButton.style.display = 'none';
+      undoButton.disabled = true;
       piling = await createMatrixPiles(matricesEl);
       history = [];
       piling.subscribe('update', updateHandler);
@@ -83,7 +85,7 @@ const createPiles = async example => {
       matricesCreditEl.style.display = 'none';
       svgEl.style.display = 'block';
       svgCreditEl.style.display = 'block';
-      undoButton.style.display = 'none';
+      undoButton.disabled = true;
       piling = await createSvgLinesPiles(svgEl);
       history = [];
       piling.subscribe('update', updateHandler);
@@ -93,6 +95,8 @@ const createPiles = async example => {
       console.warn('Unknown example:', example);
       break;
   }
+
+  return piling;
 };
 
 const exampleEl = document.getElementById('example');
@@ -126,4 +130,166 @@ switch (example) {
   // Nothing
 }
 
-createPiles(exampleEl.value);
+let isOptionsOpen = false;
+
+const handleOptionsTogglerClick = event => {
+  event.preventDefault();
+
+  isOptionsOpen = !isOptionsOpen;
+
+  if (isOptionsOpen) {
+    optionsEl.setAttribute('class', 'open');
+    document.body.setAttribute('class', `${window.example} options-open`);
+  } else {
+    optionsEl.removeAttribute('class');
+    document.body.setAttribute('class', window.example);
+  }
+};
+
+optionsTogglerEl.addEventListener('click', handleOptionsTogglerClick);
+
+createPiles(exampleEl.value).then(pilingLib => {
+  const options = [
+    {
+      id: 'grid',
+      title: 'Grid',
+      fields: [
+        {
+          name: 'itemSize',
+          dtype: 'int',
+          min: 16,
+          max: 320,
+          steps: 16,
+          nullifiable: true
+        },
+        {
+          name: 'itemPadding',
+          dtype: 'int',
+          min: 0,
+          max: 64,
+          steps: 8
+        },
+        {
+          name: 'columns',
+          dtype: 'int',
+          min: 1,
+          max: 20,
+          nullifiable: true
+        },
+        {
+          name: 'rowHeight',
+          dtype: 'int',
+          min: 16,
+          max: 320,
+          steps: 16,
+          nullifiable: true
+        },
+        {
+          name: 'cellAspectRatio',
+          dtype: 'float',
+          nullifiable: true
+        }
+      ]
+    }
+  ];
+
+  const dtypeToInputType = {
+    boolean: 'checkbox',
+    int: 'number',
+    float: 'number',
+    string: 'text'
+  };
+
+  const parseDtype = {
+    boolean: v => !!v,
+    int: v => +v,
+    float: v => +v,
+    string: v => v
+  };
+
+  const optionsContent = document.querySelector('#options .content');
+  options.forEach(section => {
+    const sectionEl = document.createElement('section');
+    sectionEl.id = section.id;
+    optionsContent.appendChild(sectionEl);
+
+    const headline = document.createElement('h4');
+    headline.textContent = section.title;
+    sectionEl.appendChild(headline);
+
+    const fields = document.createElement('div');
+    fields.setAttribute('class', 'fields');
+    sectionEl.appendChild(fields);
+
+    section.fields.forEach(field => {
+      const label = document.createElement('label');
+      const labelTitle = document.createElement('div');
+
+      const title = document.createElement('span');
+      title.setAttribute('class', 'title');
+      title.textContent = field.name;
+      labelTitle.appendChild(title);
+      label.appendChild(labelTitle);
+
+      const value = document.createElement('span');
+      value.setAttribute('class', 'value');
+      value.textContent = pilingLib.get(field.name);
+      labelTitle.appendChild(value);
+
+      const inputs = document.createElement('inputs');
+      inputs.setAttribute('class', 'inputs');
+      const input = document.createElement('input');
+      input.setAttribute('type', dtypeToInputType[field.dtype]);
+
+      if (!Number.isNaN(+field.min)) {
+        input.setAttribute('type', 'range');
+        input.setAttribute('min', +field.min);
+      }
+
+      if (!Number.isNaN(+field.max)) {
+        input.setAttribute('type', 'range');
+        input.setAttribute('max', +field.max);
+      }
+
+      if (!Number.isNaN(+field.step)) {
+        input.setAttribute('type', 'range');
+        input.setAttribute('step', +field.step);
+      }
+
+      input.setAttribute('value', pilingLib.get(field.name));
+
+      const isSet = document.createElement('input');
+      isSet.setAttribute('type', 'checkbox');
+      if (field.nullifiable) {
+        if (pilingLib.get(field.name) !== null) {
+          isSet.checked = true;
+        }
+        isSet.addEventListener('change', event => {
+          if (event.target.checked) {
+            pilingLib.set(field.name, parseDtype[field.dtype](input.value));
+            value.textContent = parseDtype[field.dtype](input.value);
+          } else {
+            pilingLib.set(field.name, null);
+            value.textContent = '';
+          }
+        });
+      } else {
+        isSet.checked = true;
+        isSet.disabled = true;
+      }
+      inputs.appendChild(isSet);
+
+      input.addEventListener('change', event => {
+        if (isSet && isSet.checked) {
+          const newValue = parseDtype[field.dtype](event.target.value);
+          pilingLib.set(field.name, newValue);
+          value.textContent = newValue;
+        }
+      });
+
+      inputs.appendChild(input);
+      label.appendChild(inputs);
+      fields.appendChild(label);
+    });
+  });
+});