+++ src/store.js
@@ -96,9 +96,13 @@ const [items, setItems] = setter('items', []);
 
 const [orderer, setOrderer] = setter('orderer', createOrderer().rowMajor);
 
-const [grid, setGrid] = setter('grid', []);
-
+// Grid
+const [itemSize, setItemSize] = setter('itemSize');
 const [itemSizeRange, setItemSizeRange] = setter('itemSizeRange', [0.7, 0.9]);
+const [columns, setColumns] = setter('columns', 10);
+const [rowHeight, setRowHeight] = setter('rowHeight');
+const [cellAspectRatio, setCellAspectRatio] = setter('cellAspectRatio', 1);
+const [itemPadding, setItemPadding] = setter('itemPadding', 0);
 
 const [itemAlignment, setItemAlignment] = setter('itemAlignment', [
   'bottom',
@@ -187,6 +191,9 @@ const [pileBackgroundOpacity, setPileBackgroundOpacity] = setter(
   'pileBackgroundOpacity',
   0.85
 );
+
+// 'topLeft', 'topRight', 'bottomLeft', 'bottomRight', 'center'
+const [pileCellAlign, setPileCellAlign] = setter('pileCellAlign', 'topLeft');
 
 const [pileContextMenuItems, setPileContextMenuItems] = setter(
   'pileContextMenuItems',
@@ -329,13 +336,17 @@ const createStore = () => {
     depiledPile,
     depileMethod,
     easing,
-    grid,
+    itemSize,
+    itemSizeRange,
+    columns,
+    rowHeight,
+    cellAspectRatio,
+    itemPadding,
     itemAlignment,
     itemOpacity,
     itemRenderer,
     itemRotated,
     items,
-    itemSizeRange,
     lassoFillColor,
     lassoFillOpacity,
     lassoStrokeColor,
@@ -351,6 +362,7 @@ const createStore = () => {
     pileBorderSize,
     pileBackgroundColor,
     pileBackgroundOpacity,
+    pileCellAlign,
     pileContextMenuItems,
     pileOpacity,
     piles,
@@ -413,8 +425,12 @@ export const createAction = {
   setCoverAggregator,
   setItems,
   setOrderer,
-  setGrid,
+  setItemSize,
   setItemSizeRange,
+  setColumns,
+  setRowHeight,
+  setCellAspectRatio,
+  setItemPadding,
   setItemAlignment,
   setItemRotated,
   setFocusedPiles,
@@ -439,5 +455,6 @@ export const createAction = {
   setPileBackgroundOpacity,
   setPileContextMenuItems,
   setPileOpacity,
-  setPileScale
+  setPileScale,
+  setPileCellAlign
 };